refactor(policy): rename policy_ext to shell_policy_ext for naming consistency
Aligns with file_policy_ext.rs naming convention per reviewer feedback.
Pure rename + reference updates, no functional changes.
